[pdb] Secondary structure implementation
This commit brings the SecondaryStructureTypes enum and the SecondaryStructure struct; these are already used by the load_pdb() and Structure

+++ bioshell-pdb/src/pdb_helix.rs
@@ -30,7 +30,7 @@ pub struct PdbHelix {
     /// Insertion code of the terminal residue
     pub end_i_code: char,
     /// Helix class
-    pub helix_class: i32,
+    pub helix_class: u8,
     /// Comment about this helix
     pub comment: String,
     /// Length of this helix
